# docs/openapi.yaml
openapi: 3.1.0
info:
  title: StateSet Sequencer API
  description: |
    Verifiable Event Sync (VES) sequencer for deterministic event ordering,
    state projection, and cryptographic commitments.

    ## Authentication

    All API endpoints require authentication via API key or JWT token, except:
    - `POST /v1/agents/register` (public self-service registration, served under the `/api` base URL)
    - `GET /health`
    - `GET /ready`
    - **API Key**: Pass in `X-API-Key` header
    - **JWT**: Pass in `Authorization: Bearer <token>` header

    ## VES Protocol

    The VES (Verifiable Event Sync) protocol provides:
    - Cryptographic signatures on events (Ed25519)
    - Gap-free sequence numbering
    - Merkle tree commitments
    - L2 anchoring for finality
    - Inclusion proofs for any event

    ## Legacy v1 Deprecation

    The pre-VES **legacy v1 surface is formally deprecated** (planned removal:
    2027-03-01): every legacy operation is marked `deprecated: true` below.
    New integrations must use the VES `/v1/ves/*` endpoints (REST) or the
    `stateset.sequencer.v2` gRPC service (`proto/sequencer_v2.proto`).

    Legacy REST -> current successor:
    - `POST /v1/events/ingest` -> `POST /v1/ves/events/ingest`
    - `GET /v1/events` -> `GET /v1/ves/events`
    - `GET /v1/head` -> `GET /v1/ves/head`
    - `/v1/commitments*` -> `/v1/ves/commitments*`
    - `/v1/proofs*` -> `/v1/ves/proofs*`
    - `/v1/entities/*` -> `/v1/ves/entities/*`
...

...
  version: 0.8.6
...
servers:
...
security:
  - apiKey: []
  - bearerAuth: []
paths:
...
  /v1/x402/payments:
    get:
      summary: List payment intents
      description: List x402 payment intents for a tenant
      operationId: listX402Payments
      tags: [x402]
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: Payment intent list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Submit payment intent
      description: Verify, sequence and persist a signed x402 payment intent
      operationId: submitX402Payment
      tags: [x402]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment intent accepted
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/x402/payments/{intent_id}:
    get:
      summary: Get payment intent
      description: Get one x402 payment intent by ID
      operationId: getX402Payment
      tags: [x402]
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment intent
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/x402/payments/{intent_id}/receipt:
    get:
      summary: Get payment receipt
      description: Get the Merkle receipt for a batched x402 payment intent
      operationId: getX402Receipt
      tags: [x402]
      parameters:
        - name: intent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment receipt
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/x402/premium/insights:
    get:
      summary: Premium insights (payment-gated)
      description: |
        Demonstration endpoint gated by the x402 payment middleware: requests
        must carry a valid signed payment intent in the `X-Payment` header or
        they receive `402 Payment Required` with the payment requirements.
      operationId: x402PremiumInsights
      tags: [x402]
      responses:
        '200':
          description: Premium insights payload
          content:
            application/json:
              schema:
                type: object
        '402':
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==========================================================================
  # Health & Metrics
  # ==========================================================================
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    tenantId:
      name: tenant_id
      in: query
      required: true
      description: Tenant identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
          description: Error type
        code:
          type: integer
          description: Numeric error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          description: Request correlation ID
        docs_url:
          type: string
          format: uri
          description: Link to error documentation
...